pyproject.toml
[project]
name = "aevis-openenv"
version = "1.0.0"
description = "AI retinal disease screening environment for OpenEnv"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
